Made both size fields required
Added the require tag to each size field to prevent the error wherein the size field is left as an empty string, which cannot be inserted into the sizes: price column, as that is a Sequelize.DECIMAL field.

Note: The required attribute of the input tag is not supported in Internet Explorer 9 and earlier versions, or in Safari. Thus, we may need to ammend this change later to account for those browsers. But honestly, fuck safari and IE.