Massive overhaul of menu GUI
A complete and total change in the way the menu GUI looks and works to avoid several graphical bugs, improve usability and visual clarity. It is also in preperation for some planned changes in the way price will be displayed and changed, prompted by the discovery of a bug.

It is missing the "change photo" button, because changes to the button on a different branch may render any work done here pointless.