Changed the way item pricing is set - no messenger changes
This changes the UI and backend to allow for price to be set individually for items and types, along with sizes. Thus, a client can add an item without a type or size, but still be able to set a price for that item. These changes have not been implemented in the messenger yet, so the end user won't see the item/type prices.