Removes price field entirely when element has child
This improves visual clarity. It wasn't this way before because the element's price was to remain even when it had childeren, and that made it a good idea to keep the element's price visible but uninteractable.... for some reason.

Now however, as the element's price is deleted when it gains a child, it makes more sense to just remove the price field entirely.