Fixed newOrder index location
newOrders are now put in the same index of the state as they would be after a page refresh. Before, this was not the case, and the orders would insert themselves in unintutive locations, reducing usability, and forcing the client to refresh in order to see the orders listed in chronological order.

Also rearranged bongNoise play call to prevent an appearance of lag, on account of drastically more calculations being proformed.